feat: add style in productitem
