--- src/components/AppRouter.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const AppRouter = () => {
-  return (
-    <div>
-      
-    </div>
-  );
-};
-
-export default AppRouter;
