+++ src/db/products.ts
@@ -1,16 +1,4 @@
-export interface Product {
-  id: number;
-  title: string;
-  description: string;
-  price: number;
-  discountPercentage: number;
-  rating: number;
-  stock: number;
-  brand: string;
-  category: string;
-  thumbnail: string;
-  images: string[];
-}
+import { Product } from 'models/product';
 
 const products: Product[] = [
   {
